Fix "While" keyword highlighting
Problem:
"while" doesn't get highlighted as a keyword

Solution:
Remove "-loop" from "While-loop"

Additional notes:
The repository also had an entry for "While-loop" in the file
ahk-mode/reference/commands, the file doesn't get included when ahk-mode is
installed (at least not from melpa), but it should probably also be changed.

"While-loop" is listed as a command in the documentation:
https://autohotkey.com/docs/commands/index.htm but its page:
https://autohotkey.com/docs/commands/While.htm only says "While-loop" in the
header, the rest of the page shows the command as "while".

And a script that contains:
while-loop () {
}

reports:
Line Text: while-loop () {
Error: This line does not contain a recognized action.